Add methods to add and validate user keys
This commit adds user key functionality to the user model.  The key
is stored hashed and salted to the users_key table, and is encoded
using a secret_key read from the config file.

The encoded key is decoded during validation, then split into it's
containing parts and looked up from the users_key table.  The
submitted key is then compared to the hashed version in the db.