Allow events to be queried by created dates
To allow searching between created dates, the events model needed
to be updated to return created_at and updated_at in unix timestamp
rather than datetime.  This commit contains updates to all of the
models.

The DateTime::Format::Strptime module was no longer
needed to verify the data in the created_at and updated_at columns,
so was removed from the deps test and README.

The events model and controller were then updated to receive the
created params, then build the queries according to the number and
definition of created params.