Update routes to flatten data return
This commit updates the routes to flatten the data return.  The
original intention was to allow multiple data structures to be
returned on request.  For example, a return might have a
collections key with events key.  That change was never made, and
frankly over complicates the API interface.

To simplify the return, a cleaner JSON datastructure is returned,
containing only the data asked for in the route, created, or
updated in the call.